refactor: create botcore package with Presenter interface and shared state
Move WizardData and state constants to internal/botcore/ so they can be
shared across platform adapters (Telegram, WhatsApp). Define the Presenter
interface that abstracts platform-specific UI rendering.

internal/bot/state.go re-exports from botcore for backward compatibility.

Closes #222
